WIP: work around cls handling of Promises
continuation-local-storage chooses to bind the context to the Promise
constructor and passes this to any and all reactions to that promise.
While this might make sense in some cases, it very much doesn't when
it comes to tracing an HTTP request context. Provide a patcher for
Promises which will bind the reaction handler to the active segment
when "then" or "catch" is called.