Create Internal Logging To Replace Winston
This commit moves from using Winston to a simplified internally
implemented logger that writes to an appropriate console method.

If there is a console method for the desired logging level then that
is used otherwise it will default to using console.log. If that
method doesn't exist then the logger will default to a noop function
meaning there will be no error but the logs will be lost.

Currently the format of the log messages has been kept very close to
the format produced by the previous code.

One change to the previous logging interface is the addition of a
way to change the level of the logger after initialisation.

The changed is based on the principal that the internal logging
should be kept as simple as possible and if the user has more
specific requirements then they should use the facility provided to
supply their own logger.