Remove SetLevel Method
This change removes the setLevel method so now only environment
variables can be used to set the log level.

It also changes how the log message is formatted so that if the
message is a falsey value it is replaced by an empty string.

Updated README to reflect the logging changes being made.

Finally a log level of silent is added which prevents all logging.